Re-organized components, add PropTypes, and ensure project structure consistency
