+++ src/components/AddTodoForm.jsx
@@ -1,6 +1,7 @@
 import { useState } from "react";
 import InputWithLabel from "./InputWithLabel";
 import styles from './TodoListItem.module.css'
+import PropTypes from "prop-types";
 
 function AddTodoForm({ onAddTodo }) {
   const [todoTitle, setTodoTitle] = useState("");
@@ -36,5 +37,9 @@ function AddTodoForm({ onAddTodo }) {
     </form>
   );
 }
+
+AddTodoForm.propTypes ={
+  onAddTodo: PropTypes.func.isRequired,
+};
 
 export default AddTodoForm;
